Major changes on build system.
User now need to generate the makefile with configure first.  It will
call autogen definition with templates to generate the good makefile
for the desired mcu.  A lot of file have been rename base on the the
file layout.  It should now be easier to add new mcu to compile and
selecting the proper devices.