Major changes in the build system
Major changes are:
      - Remove autogen
      - Remove configure script
      - Creation of directories mk/

Autogen is a major pain in the ass to use.  I've decided to ditch it
entirely.  Files that need synchronization will only use X macros now.
Not autogen.  The configuration script is also ditched for a more easy
to use system.

The mk/ directories was created to contains all makefile related
stuffs.

mk/cc.mk:
	Compiler/linker flags and options

mk/src.mk:
	Source files

mk/std.mk:
	Everything that is "standard".  Wathever that means

There's also the special file *local.mk*.  This file is use to
overload variables, such as *CFLAGS* and *VERBOSE_LEVEL*.  The
local.mk file is also use to specify the *MCU* architecture to compile
YAROS for.